speed up the matching process
Previously it was O(n^2) for n as number of lines.

Now it should be more O(nm) for m is the number of parts in a line separated by space. Normall it's 2 parts (1 space), at most I would say less than 5 (who puts space in filenames anyway!! >:( you shouldn't!).

Signed-off-by: Tin Lai <tin@tinyiu.com>